Sistemato predict e train con Pytorch

+++ utils.py
@@ -307,5 +307,5 @@ class CustomDataset(Dataset):
             'ids': torch.tensor(ids, dtype=torch.long),
             'mask': torch.tensor(mask, dtype=torch.long),
             'token_type_ids': torch.tensor(token_type_ids, dtype=torch.long),
-            'targets': torch.tensor(self.label[index], dtype=torch.float)
+            'targets': torch.tensor(self.label[index], dtype=torch.long)
         }
